Improve detection of failed builds in the mac nightly build scripts. Was still bundling and uploading broken builds :(
git-svn-id: https://rpg.hamsterrepublic.com/source/wip@6327 7d344553-34f0-0310-a9b1-970ce8f1c3a2

+++ distrib-nightly-mac.sh
@@ -38,6 +38,11 @@ cd wip
 
 ./distrib-mac.sh ${MORE_ARGS}
 
+if [ ! -f ohrrpgce-game -o ! -f ohrrpgce-custom ] ; then
+  echo Aborting distrib-nightly script because distrib script failed
+  exit 1
+fi
+
 mv distrib/OHRRPGCE-*-wip.dmg distrib/OHRRPGCE-wip.dmg
 scp -p distrib/OHRRPGCE-wip.dmg $UPLOAD_DEST/ohrrpgce/nightly/
 rm distrib/OHRRPGCE-wip.dmg
